requirements.txt
click~=8.1.8
pandas~=2.2.3
requests~=2.32.3
tqdm~=4.67.1
pyarrow~=19.0.1
cvss==3.4
GitPython~=3.1.44
strsimpy~=0.2.1
google-cloud-bigquery==3.31.0
tomli==2.2.1
fastapi[standard]~=0.115.12

# --- transitive dependencies - maybe specify versions later
numpy
cachetools
dateutil
pydantic